Feature/movie (#6)
* update gitignore

* feat[movie] : image slider add

* style : popular movie list layout design

* feat[movie] : dataService & model class add

* feat : movie list demo service add

* refactor : code indentation

* feat[movie] : Dashboard popular movie view implementation

* feat[movie] : toprated & nowplaying movie list service & ui implementation

* fix[movie] : scrolllview position problem

* feat[movie] : recycler item decoration add

* feat[movie] : item decorator dependency injection add

* feat[movie] : upcoming movie service & ui implementation

* refactor[movie] : Image slider

* feat[movie] : update slider item

* feat[movie] : movie list paging 3 setup

* feat[movie] : List paginag implementation

* refactor[movie] : movie list

* feat[movie] : see all click & sageArgs implementation

* feat[movie] : list fragment toolbar title set with category

* details & collection page done (#4)

movie details & collection done

* refactor[movie] : movie slider onclick implementation

* feat[firebase] : firebase add

* feat[movie] : Slide movie name add

* feat[movie] : movie icon add

* feat[app] : shared module res support for dark mode

+++ movie/src/main/java/com/saiful/movie/model/MovieDetails.kt
@@ -14,7 +14,7 @@ data class MovieDetails(
 	val title: String? = null,
 	@field:Json(name="backdrop_path")
 	val backdropPath: String? = null,
-	val revenue: Int? = null,
+	val revenue: Double? = null,
 	val genres: List<GenresItem?>? = null,
 	val popularity: Double? = null,
 	@field:Json(name="production_countries")
